Swap out the salary negotiation post for a recent favorite.
The salary negotiation post is still inspiring, but I've come to learn that it doesn't apply in many situations, like if you are founding a company, working on an early stage startup, working in zero-sum salary situation, or freelancing. Also, the other Kalzumeus post briefly addresses some negotiating basics that are sufficient for now.